+++ src/Hashing/Hasher.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Sentinel
- * @version    2.1.0
+ * @version    2.0.11
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -36,9 +36,13 @@ trait Hasher
      */
     protected function createSalt()
     {
-        $pool = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-
-        return substr(str_shuffle(str_repeat($pool, 5)), 0, $this->saltLength);
+        $pool = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ./';
+        $max = strlen($pool) - 1;
+        $salt = '';
+        for ($i = 0; $i < $this->saltLength; ++$i) {
+            $salt .= $pool[random_int(0, $max)];
+        }
+        return $salt;
     }
 
     /**